feat: remove callback and do not generate doc

+++ oxidros-msg/src/primitives.rs
@@ -275,22 +275,20 @@ impl<const N: usize> RosString<N> {
         Self::assign_string(&mut self.0, s)
     }
 
-    #[allow(clippy::unnecessary_cast)]
     pub fn as_slice(&self) -> &[std::os::raw::c_char] {
         if self.0.data.is_null() {
             &[]
         } else {
-            let s = unsafe { std::slice::from_raw_parts(self.0.data, self.0.size as usize) };
+            let s = unsafe { std::slice::from_raw_parts(self.0.data, self.0.size) };
             s
         }
     }
 
-    #[allow(clippy::unnecessary_cast)]
     pub fn as_slice_mut(&mut self) -> &mut [std::os::raw::c_char] {
         if self.0.data.is_null() {
             &mut []
         } else {
-            let s = unsafe { std::slice::from_raw_parts_mut(self.0.data, self.0.size as usize) };
+            let s = unsafe { std::slice::from_raw_parts_mut(self.0.data, self.0.size) };
             s
         }
     }
@@ -359,12 +357,11 @@ impl<const STRLEN: usize, const SEQLEN: usize> RosStringSeq<STRLEN, SEQLEN> {
         Self(msg)
     }
 
-    #[allow(clippy::unnecessary_cast)]
     pub fn as_slice(&self) -> &[RosString<STRLEN>] {
         if self.0.data.is_null() {
             &[]
         } else {
-            let s = unsafe { std::slice::from_raw_parts(self.0.data, self.0.size as usize) };
+            let s = unsafe { std::slice::from_raw_parts(self.0.data, self.0.size) };
             unsafe { transmute::<&[rosidl_runtime_c__String], &[RosString<STRLEN>]>(s) }
         }
     }
